src/requirements.txt
cloudshell-shell-core>=4.0.0,<4.1.0
cloudshell-networking>=5.1.0,<5.2.0
cloudshell-networking-devices>=2.7.0,<2.8.0
cloudshell-snmp<=3.3.0,<3.4.0